Test for jpgs vs pdfs

+++ php/loadImage.php
@@ -9,7 +9,7 @@ if( array_key_exists( 'id', $_REQUEST ) ){
 if( array_key_exists( 'size', $_REQUEST ) ) {
 	$size =  $_REQUEST['size'];
 } else {
-	//return
+	$size = 'full';
 }
 
 if( array_key_exists( 'page', $_REQUEST ) ) {
@@ -19,34 +19,46 @@ if( array_key_exists( 'page', $_REQUEST ) ) {
 }
 
 //development vars
-$id = 37;
+$id = 72;
 $size = 'full';
 $page = 1;
 
 if( ! file_exists( getcwd(). '\images\\' . $id . '-' . $size . '-' . $page . '.jpg' ) ){
-	file_put_contents( 'temp.pdf', file_get_contents( 'http://cdm15963.contentdm.oclc.org/utils/getfile/collection/p15963coll18/id/' . $id ) );
-
-	//check amount of pages in pdf
-	$im = new Imagick();
-	$im->readimage( getcwd() . '\temp.pdf' );
-	$pages = $im->getNumberImages();
-	$im->clear();
-	$im->destroy();
-
-	//convert pdf page to jpg
-	$imagick = new Imagick();
-	$imagick->setResolution(300,300);
-
-	$imagick->readimage( getcwd() . '\temp.pdf[' . $page . ']' );
-	$imagick->setImageFormat( 'jpeg' );
-	$imagick->writeImage( getcwd() . '\images\\' . $id . '-' . $size . '-' . $page . '.jpg' );
-		
-	$imagick->clear();
-	$imagick->destroy();
+	$file_info = json_decode( file_get_contents( 'http://server15963.contentdm.oclc.org/dmwebservices/index.php?q=dmGetItemInfo/p15963coll18/' . $id . '/json' ) );
 	
-	unlink( "temp.pdf" );
+	//file is a pdf
+	if( substr( $file_info->{'find'}, -3 ) == 'pdf' ) {
+	
+		file_put_contents( 'temp.pdf', file_get_contents( 'http://cdm15963.contentdm.oclc.org/utils/getfile/collection/p15963coll18/id/' . $id ) );
+
+		//check amount of pages in pdf
+		$im = new Imagick();
+		$im->readimage( getcwd() . '\temp.pdf' );
+		$pages = $im->getNumberImages();
+		$im->clear();
+		$im->destroy();
+
+		//convert pdf page to jpg
+		$imagick = new Imagick();
+		$imagick->setResolution(300,300);
+		
+		$i = 0;
+		while( $i <= $pages - 1 ){
+			$imagick->readimage( getcwd() . '\temp.pdf[' . $i . ']' );
+			$imagick->setImageFormat( 'jpeg' );
+			$imagick->writeImage( getcwd() . '\images\\' . $id . '-' . $size . '-' . $i . '.jpg' );
+			$i++;
+		}
+			
+		$imagick->clear();
+		$imagick->destroy();
+		
+		unlink( "temp.pdf" );
+	} elseif ( substr( $file_info->{'find'}, -3 ) == 'jpg' OR substr( $file_info->{'find'}, -4 ) == 'jpeg' ){
+		echo 'This is a jpg';
+	}
 }
 
-// header( "Location: images/$id-0.jpg" );
+// header( "Location: images/$id-$size-$page.jpg" );
 
 ?>
